+++ lib/transaction/utils.go
@@ -5,10 +5,9 @@ import (
 	"fmt"
 	"log"
 	"net/http"
-	"strconv"
-	"strings"
-	"time"
+	"sort"
 
+	"github.com/btcsuite/btcd/btcjson"
 	"github.com/btcsuite/btcd/btcutil"
 	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcd/wire"
@@ -44,6 +43,50 @@ func isSegWitAddress(addr btcutil.Address) bool {
 	default:
 		return false
 	}
+}
+
+func selectAdditionalUTXOs(w *wallet.Wallet, amount btcutil.Amount) ([]*btcjson.ListUnspentResult, btcutil.Amount, error) {
+	utxos, err := w.ListUnspent(1, 9999999, "")
+	if err != nil {
+		return nil, 0, fmt.Errorf("failed to list unspent outputs: %v", err)
+	}
+
+	var selectedUTXOs []*btcjson.ListUnspentResult
+	var totalSelected btcutil.Amount
+
+	// Sort UTXOs by amount in descending order
+	sort.Slice(utxos, func(i, j int) bool {
+		return utxos[i].Amount > utxos[j].Amount
+	})
+
+	for _, utxo := range utxos {
+		if totalSelected >= amount {
+			break
+		}
+		selectedUTXOs = append(selectedUTXOs, utxo)
+		totalSelected += btcutil.Amount(utxo.Amount * btcutil.SatoshiPerBitcoin)
+	}
+
+	if totalSelected < amount {
+		return nil, 0, fmt.Errorf("insufficient funds to cover additional fee")
+	}
+
+	return selectedUTXOs, totalSelected, nil
+}
+
+func fetchUTXO(w *wallet.Wallet, outpoint *wire.OutPoint) (*btcjson.ListUnspentResult, error) {
+	utxos, err := w.ListUnspent(1, 9999999, "")
+	if err != nil {
+		return nil, fmt.Errorf("failed to list unspent outputs: %v", err)
+	}
+
+	for _, utxo := range utxos {
+		if utxo.TxID == outpoint.Hash.String() && utxo.Vout == outpoint.Index {
+			return utxo, nil
+		}
+	}
+
+	return nil, fmt.Errorf("UTXO not found")
 }
 
 func UpdateUTXOs(w *wallet.Wallet) error {
@@ -90,51 +133,6 @@ func VerifyUTXO(txID string, vout uint32) error {
 	}
 
 	return nil
-}
-
-func getFeeRecommendation() (FeeRecommendation, error) {
-	client := &http.Client{Timeout: 10 * time.Second}
-	resp, err := client.Get("https://mempool.space/api/v1/fees/recommended")
-	if err != nil {
-		return FeeRecommendation{}, err
-	}
-	defer resp.Body.Close()
-
-	var feeRec FeeRecommendation
-	err = json.NewDecoder(resp.Body).Decode(&feeRec)
-	return feeRec, err
-}
-
-func getUserFeePriority(feeRec FeeRecommendation) (int, error) {
-	fmt.Println("Choose your fee priority:")
-	fmt.Printf("1. Fastest (%.2f sat/vB)\n", float64(feeRec.FastestFee))
-	fmt.Printf("2. Half Hour (%.2f sat/vB)\n", float64(feeRec.HalfHourFee))
-	fmt.Printf("3. Hour (%.2f sat/vB)\n", float64(feeRec.HourFee))
-	fmt.Printf("4. Economy (%.2f sat/vB)\n", float64(feeRec.EconomyFee))
-	fmt.Printf("5. Minimum (%.2f sat/vB)\n", float64(feeRec.MinimumFee))
-	fmt.Print("Enter your choice (1-5): ")
-
-	var input string
-	fmt.Scanln(&input)
-	choice, err := strconv.Atoi(strings.TrimSpace(input))
-	if err != nil || choice < 1 || choice > 5 {
-		return 0, fmt.Errorf("invalid choice")
-	}
-
-	switch choice {
-	case 1:
-		return feeRec.FastestFee, nil
-	case 2:
-		return feeRec.HalfHourFee, nil
-	case 3:
-		return feeRec.HourFee, nil
-	case 4:
-		return feeRec.EconomyFee, nil
-	case 5:
-		return feeRec.MinimumFee, nil
-	default:
-		return 0, fmt.Errorf("unexpected error")
-	}
 }
 
 func findUnusedChangeAddress(w *wallet.Wallet) (btcutil.Address, error) {
