--- lib/types/types.go
@@ -1,6 +1,0 @@
-package types
-
-type Config struct {
-	BackendURL string   `mapstructure:"backend_url"`
-	AddPeers   []string `mapstructure:"add_peers"`
-}
